Pay Period interaction  + Search Button
I was able to make the Pay Period button interactable even though it was dynamic in terms of time. Was told by Arssh I don't need to do the backend for this and I just need to make a search button that is clickable and wires back to the backend. After that I was able to spend some time to perfectly align it